pyproject.toml
[project]
name = "fileManagerMcpServer"
version = "1.0.0"
description = "Advanced MCP server for comprehensive file management operations"
authors = [
    {name = "Nima M", author= "https://github.com/mlnima"}
]
dependencies = [
    "fastmcp>=0.1.0",
]
requires-python = ">=3.8"
readme = "README.md"
license = {text = "MIT"}

[project.urls]
Homepage = "https://02dev.com"
Repository = "https://github.com/mlnima"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project.scripts]
file-manager-mcp = "server:main"